fixed build flow (not finished)

+++ app/src/main/java/com/superbgoal/caritasrig/data/model/Memory.kt
@@ -3,10 +3,10 @@ package com.superbgoal.caritasrig.data.model
 data class Memory(
     val name: String,
     val price: Double,
-    val speed: List<Int>, // List of Int for speed values (e.g., [4, 3200])
-    val modules: List<Int>, // List of Int for module sizes (e.g., [2, 8])
-    val pricePerGb: Double,
+    val speed: List<Double>, // List of Double for speed values (e.g., [5, 6600])
+    val modules: List<Int>, // List of Int for module sizes (e.g., [2, 16])
+    val pricePerGb: Double, // Renamed from price_per_gb for camelCase
     val color: String,
-    val firstWordLatency: Int, // First word latency in nanoseconds
-    val casLatency: Int // CAS latency
+    val firstWordLatency: Double, // Updated to Double for correct mapping of decimal values
+    val casLatency: Int // CAS latency remains as Int
 )
